Trait jf_merkle_tree::namespaced_merkle_tree::BindNamespace

pub trait BindNamespace<E, I, T, N>: DigestAlgorithm<E, I, T>
where
    E: Element,
    N: Namespace,
    T: NodeValue,
    I: Index,
{
    // Required method
    fn generate_namespaced_commitment(
        namespaced_hash: NamespacedHash<T, N>,
    ) -> T;
}

Trait indicating that a digest algorithm can commit to a namespace range.

Required Methods

fn generate_namespaced_commitment(namespaced_hash: NamespacedHash<T, N>) -> T

Generate a commitment that binds a node to a namespace range

Object Safety

This trait is not object safe.