jf_merkle_tree

Trait DigestAlgorithm

pub trait DigestAlgorithm<E, I, T>
where
    E: Element,
    I: Index,
    T: NodeValue,
{
    // Required methods
    fn digest(data: &[T]) -> Result<T, MerkleTreeError>;
    fn digest_leaf(pos: &I, elem: &E) -> Result<T, MerkleTreeError>;
}

Merkle tree hash function WARN: it’s important to domain separate the two digest functions. Otherwise, you may suffer from the length extension attack.

Required Methods

fn digest(data: &[T]) -> Result<T, MerkleTreeError>

Digest a list of values

fn digest_leaf(pos: &I, elem: &E) -> Result<T, MerkleTreeError>

Digest an indexed element

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.