Trait jf_merkle_tree::MerkleTreeScheme

pub trait MerkleTreeScheme: Sized {
    type Element: Element;
    type Index: Index;
    type NodeValue: NodeValue;
    type MembershipProof: Clone + Eq + Hash;
    type BatchMembershipProof: Clone;
    type Commitment: MerkleCommitment<Self::NodeValue>;

    const ARITY: usize;

    // Required methods
    fn height(&self) -> usize;
    fn capacity(&self) -> BigUint;
    fn num_leaves(&self) -> u64;
    fn commitment(&self) -> Self::Commitment;
    fn lookup(
        &self,
        pos: impl Borrow<Self::Index>
    ) -> LookupResult<&Self::Element, Self::MembershipProof, ()>;
    fn verify(
        root: impl Borrow<Self::NodeValue>,
        pos: impl Borrow<Self::Index>,
        proof: impl Borrow<Self::MembershipProof>
    ) -> Result<Result<(), ()>, MerkleTreeError>;
    fn iter(
        &self
    ) -> MerkleTreeIter<'_, Self::Element, Self::Index, Self::NodeValue>;
}

Basic functionalities for a merkle tree implementation. Abstracted as an accumulator for fixed-length array. Supports generate membership proof at a given position and verify a membership proof.

Required Associated Types

type Element: Element

Merkle tree element type

type Index: Index

Index type for this merkle tree

type NodeValue: NodeValue

Internal and root node value

type MembershipProof: Clone + Eq + Hash

Merkle proof

type BatchMembershipProof: Clone

Batch proof

type Commitment: MerkleCommitment<Self::NodeValue>

Merkle tree commitment

Required Associated Constants

const ARITY: usize

Tree ARITY

Required Methods

fn height(&self) -> usize

Return the height of this merkle tree

fn capacity(&self) -> BigUint

Return the maximum allowed number leaves

fn num_leaves(&self) -> u64

Return the current number of leaves

fn commitment(&self) -> Self::Commitment

Return a merkle commitment

fn lookup( &self, pos: impl Borrow<Self::Index> ) -> LookupResult<&Self::Element, Self::MembershipProof, ()>

Returns the leaf value given a position

pos - zero-based index of the leaf in the tree

returns - Leaf value at the position along with a proof. LookupResult::EmptyLeaf if the leaf position is empty or invalid, LookupResult::NotInMemory if the leaf position has been forgotten.

fn verify( root: impl Borrow<Self::NodeValue>, pos: impl Borrow<Self::Index>, proof: impl Borrow<Self::MembershipProof> ) -> Result<Result<(), ()>, MerkleTreeError>

Verify an element is a leaf of a Merkle tree given the proof

root - a merkle tree root, usually obtained from Self::commitment().digest()

pos - zero-based index of the leaf in the tree

proof - a merkle tree proof

returns - Ok(true) if the proof is accepted, Ok(false) if not. Err() if the proof is not well structured, E.g. not for this merkle tree.

fn iter( &self ) -> MerkleTreeIter<'_, Self::Element, Self::Index, Self::NodeValue>

Return an iterator that iterates through all element that are not forgetton

Object Safety

This trait is not object safe.