jf_merkle_tree

Trait RangeProofMerkleTreeScheme

pub trait RangeProofMerkleTreeScheme: MerkleTreeScheme {
    type RangeMembershipProof;

    // Required methods
    fn range_lookup(
        &self,
        start: impl Borrow<Self::Index>,
        end: impl Borrow<Self::Index>,
    ) -> LookupResult<(Vec<Self::Index>, Vec<Self::Element>), Self::RangeMembershipProof, ()>;
    fn verify_range_proof(
        commitment: impl Borrow<Self::Commitment>,
        indices: &[impl Borrow<Self::Index>],
        elements: &[impl Borrow<Self::Element>],
        proof: impl Borrow<Self::RangeMembershipProof>,
    ) -> Result<Result<(), ()>, MerkleTreeError>;
}

Merkle tree that supports generating range proofs

Required Associated Types

type RangeMembershipProof

Range proof for a given range

Required Methods

fn range_lookup( &self, start: impl Borrow<Self::Index>, end: impl Borrow<Self::Index>, ) -> LookupResult<(Vec<Self::Index>, Vec<Self::Element>), Self::RangeMembershipProof, ()>

Generate a range proof for leaves in [start, end], inclusive.

start - zero-based start index of the range (inclusive)

end - zero-based end index of the range (inclusive)

returns - Indexed leaf values in the range along with a range membership proof.

NOTE: this interface will return NotInMemory or NotFound if any of the leaves in the range are not in memory or not found.

fn verify_range_proof( commitment: impl Borrow<Self::Commitment>, indices: &[impl Borrow<Self::Index>], elements: &[impl Borrow<Self::Element>], proof: impl Borrow<Self::RangeMembershipProof>, ) -> Result<Result<(), ()>, MerkleTreeError>

Verify a range proof for leaves in [start, end], inclusive.

commitment - a merkle tree commitment

indices - zero-based indices of the leaves in the tree

elements - the leaf values in the range

proof - a range membership proof for the given range

returns - Ok(SUCCESS) if the proof is accepted, Ok(FAIL) if not. Err() if the proof is not well structured, E.g. not for this merkle tree.

NOTE: all indices are needed here because we cannot generate them from [start, end] under the current Index trait bound.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.