Trait jf_merkle_tree::hasher::HasherDigest

pub trait HasherDigest: Digest<OutputSize = Self::Foo> + Write + Send + Sync {
    type Foo: ArrayLength<u8, ArrayType = Self::Bar>;
    type Bar: Copy;
}

Convenience trait and blanket impl for downstream trait bounds.

Useful for downstream code that’s generic offer [Digest] hasher H.

Example

Do this:

fn generic_over_hasher<H>()
where
    H: HasherDigest,
{
    let my_data = [1, 2, 3, 4, 5, 6, 7, 8, 9];
    let mt = HasherMerkleTree::<H, usize>::from_elems(None, &my_data).unwrap();
}

Instead of this:

fn generic_over_hasher<H>()
where
    H: Digest + Write + Send + Sync,
    <<H as OutputSizeUser>::OutputSize as ArrayLength<u8>>::ArrayType: Copy,
{
    let my_data = [1, 2, 3, 4, 5, 6, 7, 8, 9];
    let mt = HasherMerkleTree::<H, usize>::from_elems(None, &my_data).unwrap();
}

Note that the complex trait bound for Copy is necessary:

fn generic_over_hasher<H>()
where
    H: Digest + Write + Send + Sync,
{
    let my_data = [1, 2, 3, 4, 5, 6, 7, 8, 9];
    let mt = HasherMerkleTree::<H, usize>::from_elems(None, &my_data).unwrap();
}

Required Associated Types

type Foo: ArrayLength<u8, ArrayType = Self::Bar>

Associated type needed to express trait bounds.

type Bar: Copy

Associated type needed to express trait bounds.

Object Safety

This trait is not object safe.